requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
requests==2.32.5
urllib3==2.6.3
python-telegram-bot==22.5
pymongo==4.16.0
dnspython==2.8.0

#python-telegram-bot==13.15